% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gamma0.R
\name{compute_gamma0}
\alias{compute_gamma0}
\title{Compute gamma0, given haplotype frequencies}
\usage{
compute_gamma0(N0, N1, W, gamma.CV, freq)
}
\arguments{
\item{N0}{number of samples with Y=0}

\item{N1}{number of samples with Y=1}

\item{W}{The causal SNPs}

\item{gamma.CV}{The log odds ratios of effect for each CV}

\item{freq}{Frequencies of SNP appearances (computed using snphap)}
}
\value{
The value of gamma0
}
\description{
Compute the value of gamma_0 from gamma1,...,gamma_m
}
\details{
Note: assume we must compute the distibution of W in the controls
}
\author{
Mary Fortune
}

