% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_zscore.R
\name{est_statistic}
\alias{est_statistic}
\title{estimate Z score at a single SNP}
\usage{
est_statistic(N0, N1, snps, W, gamma1, freq, GenoProbList)
}
\arguments{
\item{N0}{The number of Y=0}

\item{N1}{The number of Y=1}

\item{snps}{The snps at which we wish to compute the expected Z Score}

\item{W}{The true causal SNPs (these need not be in "snps")}

\item{gamma1}{The odds ratios of effect of the true causal SNPs (not including gamma0, the intercept term)}

\item{freq}{Frequencies of SNP appearances (computed using snphap)}

\item{GenoProbList}{An list of objects giving the probability of seeing each {X,W} genotype vector}
}
\value{
The expected Z Score for SNP X, assuming the causal SNPs are W
}
\author{
Mary Fortune and Chris Wallace
}
