% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_zscore.R
\name{est_zscore}
\alias{est_zscore}
\title{estimate Z score at a single SNP}
\usage{
est_zscore(N0, N1, Ufactor, powerfactor, freq, GenoProbXW)
}
\arguments{
\item{N0}{The number of Y=0}

\item{N1}{The number of Y=1}

\item{Ufactor}{The constant factor used to compute the expectation of U}

\item{powerfactor}{The constant factor used to compute the expectation of the genotype of X to some power}

\item{freq}{Frequencies of SNP appearances (computed using snphap)}

\item{GenoProbXW}{An object giving the probability of seeing each {X,W} genotype vector}
}
\value{
The expected Z Score for SNP X, assuming the causal SNPs are W
}
\description{
Estimates the expected Z Score for a single SNP, assuming the input CVs, and the relationship, gamma, between them and the trait of interest
Assumes we have already generated GenoProbXW for all X
}
\author{
Mary Fortune
}

