% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGWAS.R
\name{expected_z_score}
\alias{expected_z_score}
\title{Compute expected Z Score}
\usage{
expected_z_score(N0, N1, snps, W, gamma.CV, freq)
}
\arguments{
\item{N0}{The number of Y=0}

\item{N1}{The number of Y=1}

\item{snps}{The snps at which we wish to compute the expected Z Score}

\item{W}{The true causal SNPs (these need not be in "snps")}

\item{gamma.CV}{The log odds ratios of effect of the true causal SNPs}

\item{freq}{Frequencies of SNP appearances (computed using snphap)}
}
\value{
The expected Z Score for all snps in snps, assuming the causal SNPs are W
}
\author{
Mary Fortune
}

