% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{haps}
\alias{haps}
\title{simGWAS_test_data}
\format{
A matrix of haplotypes, one row per haplotype, one column per SNP
}
\usage{
data(simGWAS_test_data)
}
\description{
Simulated data to use in testing and vignettes in the coloc package
}
\examples{
data(simGWAS_test_data) # loads object called haps
str(haps)
}
\keyword{datasets}
