% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dataset.R
\name{make_dataset}
\alias{make_dataset}
\title{Make simulated data which has N0 cases, N1 controls and is consitent with the causal model {W,gamma}}
\usage{
make_dataset(df, N0, N1, CV, PWgY0, PWgY1)
}
\arguments{
\item{df}{dataset of genotypes to sample from (the control samples)}

\item{N0}{number of samples with Y=0}

\item{N1}{number of samples with Y=1}

\item{CV}{which SNPs are causal?}

\item{PWgY0}{the distribution of causal snps in controls}

\item{PWgY1}{the distribution of causal snps in cases}
}
\value{
simulated genotype data for N0 controls and N1 cases
}
\author{
Mary Fortune
}

