% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGWAS.R
\name{simulated_z_score}
\alias{simulated_z_score}
\title{Compute a simulated Z Score}
\usage{
simulated_z_score(N0, N1, snps, W, gamma.CV, freq, df_control)
}
\arguments{
\item{N0}{The number of Y=0}

\item{N1}{The number of Y=1}

\item{snps}{The snps at which we wish to compute the expected Z Score}

\item{W}{The true causal SNPs (these need not be in "snps")}

\item{gamma.CV}{The odds ratios of effect of the true causal SNPs}

\item{freq}{Frequencies of SNP appearances (computed using snphap)}

\item{df_control}{A reference set of control samples}
}
\author{
Mary Fortune
}

