% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractsnps.R
\name{which_genotypes}
\alias{which_genotypes}
\title{computes a matrix telling us which haplotype pairs correspond to which genotypes}
\usage{
which_genotypes(nsnps)
}
\arguments{
\item{nsnps}{The number of SNPs}
}
\author{
Mary Fortune
}
