% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_GenoProbList.R
\name{make_GenoProbList}
\alias{make_GenoProbList}
\title{make_GenoProbList}
\usage{
make_GenoProbList(snps, W, freq)
}
\arguments{
\item{snps}{Index of SNPs at which we wish to compute the expected Z Score}

\item{W}{Index of true causal SNPs (these need not be in "snps")}

\item{freq}{Frequencies of SNP appearances (computed using snphap)}
}
\value{
The the GenoProb values for each X
}
\description{
compute a list, GenoProbList, giving the GenoProb values for each X.
}
\examples{
freq=fake_freq() # fake haplotype frequency data
problist=make_GenoProbList(1:2,W=1,freq)
}
\author{
Mary Fortune and Chris Wallace
}
